Track hiding, perpetuity and no more tweepy.
Tracks can now be hidden and, when created with the XML upload form, are
hidden by default. After uploading, the authenticated user is taken to a
list of all hidden tracks for review. Track hiding and unhiding appears
in the admin selection menu, as well as in each track's actions thing.
Playing a track unhides it. Closes #18 and addresses an issue I'm
surprised to see isn't actually in the issue tracker.

Tweepy was misbehaving on now playing tweets, so both I'm using twitter
for everything now. Distressingly, twitter looks like it's even less
likely to be updated to API 1.1.

Other behaviour changes include that tracks can now be marked as played
even if they're ineligible (unless they've already been played this
week), more controls appear on ineligible tracks, a fix for titles not
appearing on tracks with extremely short titles (like that one 765PRO
ALLSTARS track ‘i’).