Break everything, but make the model APIs nicer (and more normalised).
Notable exceptions to normalisation include some objects which store
both a time of creation and the show to which they apply. I'm way more
fine with this than with the current, worse evil of it being dog-slow to
work out what show a play or vote or whatever is for.

The primary source of additional normalisation is in giving Twitter
users their own database object. This, as a bonus, will also allow a
very easy fix for the broken user images that have become super obvious
in the last couple of months as people have been regularly changing
theirs.

I probably missed some things, too. We'll get there.