Link to the shiny new canonical Neko Desu homepage.
Getting rid of the irc/webchat links will have to wait for
NekomimiScience/NekoDesu.co.uk#1.

*glares*

Also, It'd be cool to get the neko desu link out of the nav bar and into
the header somewhere, but I'm not sure entirely where I'd put it. I
suppose the 'taking votes for'/'listen now' thing doesn't really belong
where it is now anyway, so the real question is probably where to put
*that*.