Don't throw an error when looking for a closure that doesn't exist.
This appears to be an implementation detail that changed in Python 3.
Perhaps it's documented and now there's a better way to do this than
looking for an object with a `section` attribute, but right now I'm in
the business of doing the bare minimum to make this site work again.

That said, the login view is still 500ing, so I'm probably not doing a
very good job.