Implement decade roulette.
Doesn't quite look like the screenshot, but closes #193.

Sorry, @homsar, I missed your comment about the short track roulette
stuff until I was writing this commit message, but it seems like we did
this in pretty similar ways, so I'm happy to do a fun merge with yours
and push that soon.