Don't trip up over Linode's DDoS protection.
I feel like a privacy policy probably shouldn't be behind one of these,
but there's not a lot other than this I can do about it. I'm not about
to fire up selenium sessions to test other people's websites.

I do think it's unlikely that Linode will change this URL, though, so
we'll probably be fine.