Stop mentioning twitter in the README.
We had talked about mentioning what has changed as a result of #235
here, but on reflection, I don't think that makes much sense. Existing
users are going to behave habitually, and we have placed a net to catch
them in the form of the login page. There's no reason to mention any of
this stuff to new users.