Be very explicit about why we're asking people to disconnect.
I think opening with this is a good idea, just to make sure we're on the
same page as the reader. Without this context, it might be possible for
someone to come away being unsure why we want them to do this, and I can
imagine that being a pretty confusing experience.