Have git and flake8 common virtualenv directories.
There's an argument to be made that this should be configured per-user
in their global gitignore, but since we're specifically recommending a
tutorial that recommends you make a virtualenv called `venv` in the
working directory, it's probably reasonable to exclude them here too.

I've also asked flake8 to ignore node_modules. While we're not using
any, I think there are npm packages that contain Python code.

This will hopefully make invoking flake8 in a commit hook easier for
people who keep their virtualenvs in-path.

cc. #240